requirements.txt
alembic==1.6.5
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
flask_cors==3.0.10
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.1
psycopg2-binary==2.9.1
python-dateutil==2.8.1
python-dotenv==0.19.0
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.18
Werkzeug==2.0.1
WTForms==2.3.3

